Improvements & changes after review
