+++ webpack.config.js
@@ -31,7 +31,7 @@ module.exports = function makeConfig(grunt) {
     return {
         entry: {
             app: [
-                path.join(__dirname, 'scripts', 'translations.js'),
+                path.join(__dirname, 'scripts', 'translations'),
                 path.join(__dirname, 'scripts', 'index'),
             ],
         },
